+++ extensions-core/parquet-extensions/src/main/java/org/apache/druid/data/input/parquet/DruidParquetAvroInputFormat.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.druid.data.input.parquet.avro;
+package org.apache.druid.data.input.parquet;
 
 import org.apache.avro.generic.GenericRecord;
 import org.apache.parquet.avro.DruidParquetAvroReadSupport;
